Correlated subqueries!
I don't much like that I'm using the Storage object to handle stack frames, and that I have to reattach storage objects to table references based on the stack frame; but this works for now.

I likely should rewrite this such that table objects are instead middleware pieces that stack data throughout the pipeline.